Representation of generic objects matches

Matches between query and target generic objects can be represented by the Matched object. By default, all data accessors work as left joins between the query and the target object, i.e. values are returned for each query object with eventual duplicated entries (values) if the query object matches more than one target object. See also Creation and subsetting as well as Extracting data sections below for details and more information

Matched(
  query = list(),
  target = list(),
  matches = data.frame(query_idx = integer(), target_idx = integer(), score =
    numeric())
)

# S4 method for Matched
length(x)

# S4 method for Matched
show(object)

# S4 method for Matched
[(x, i, j, ..., drop = FALSE)

target(object)

query(object)

whichTarget(object)

whichQuery(object)

# S4 method for Matched
$(x, name)

# S4 method for Matched
colnames(x)

matchedData(object, ...)

# S4 method for Matched
matchedData(object, columns = colnames(object), ...)

pruneTarget(object)

Arguments

query
object with the query elements.

target
object with the elements against which query has been matched.

matches
data.frame with columns "query_idx" (integer), "target_idx" (integer) and "score" (numeric) representing the n:m mapping of elements between the query and the target objects.

x
Matched object.

object
a Matched object.

i
integer or logical defining the query elements to keep.

j
for [: ignored.

...
additional parameters.

drop
for [: ignored.

name
for $: the name of the column (or variable) to extract.

columns
for matchedData: character vector with column names of variables that should be extracted.

Value

See individual method description above for details.

Creation and subsetting

Matched objects can be created with the Matched function providing the query and target object as well as the matches data.frame with two columns of integer indices defining which elements from query match which element from target.

A Matched object can be subsetted in the following ways:

[ subset the Matched selecting query object elements to keep with parameter i. The resulting Matched object will contain all the matches for the selected query elements. The target object will by default be returned as-is.

pruneTarget cleans the Matched object by removing non-matched target elements.

Extracting data

$ extracts a single variable from the Matched x. The variables that can be extracted can be listed using colnames(x). These variables can belong to query, target or be related to the matches (e.g. the score of each match). If the query (target) object is two dimensional, its columns can be extracted (prefix "target_" is used for columns in the target object) otherwise (if query respectively target does have only a single dimension (e.g. is a list or a character) the whole object can is extracted with x$query (x$target). The matching scores are available as variable "score". Similar to a left join between the query and target elements, this function returns a value for each query element, with eventual duplicated values for query elements matching more than one target element. If variables from the target data.frame are extracted, an NA is reported for the entries corresponding to query elements that don't match any target element. See examples below for more details.

length returns the number of query elements.

matchedData allows to extract multiple variables contained in the Matched object as a DataFrame. Parameter columns allows to define which columns (or variables) should be returned (defaults to columns = colnames(object)). Each single column in the returned DataFrame is constructed in the same way as in $. That is, like $, this function performs a left join of variables from the query and target objects returning all values for all query elements (eventually returning duplicated elements for query elements matching multiple target elements) and the values for the target elements matched to the respective query elements (or NA if the target element is not matched to any query element).

target returns the target object.

query returns the query object.

whichTarget returns an integer with the indices of the elements in target that match at least one element in query.

whichQuery returns an integer with the indices of the elements in query that match at least one element in target.

Author

Andrea Vicini, Johannes Rainer

Examples

## Creating a Matched object.
q1 <- data.frame(col1 = 1:5, col2 = 6:10)
t1 <- data.frame(col1 = 11:16, col2 = 17:22)
## Define matches between query row 1 with target row 2 and, query row 2
## with target rows 2,3,4 and query row 5 with target row 5.
mo <- Matched(
    q1, t1, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
                                target_idx = c(2L, 2L, 3L, 4L, 5L),
                                 score = seq(0.5, 0.9, by = 0.1)))
mo
#> Object of class Matched 
#> Total number of matches: 5 
#> Number of query objects: 5 (3 matched)
#> Number of target objects: 6 (4 matched)

## Which of the query elements (rows) match at least one target
## element (row)?
whichQuery(mo)
#> [1] 1 2 5

## Which target elements (rows) match at least one query element (row)?
whichTarget(mo)
#> [1] 2 3 4 5

## Extracting variable "col1" from query object .
mo$col1
#> [1] 1 2 2 2 3 4 5

## We have duplicated values for the entries of `col1` related to query
## elements (rows) matched to multiple rows of the target object). The
## value of `col1` is returned for each element (row) in the query.

## Extracting variable "col1" from target object. To access columns from
## target we have to prefix the name of the column by `"target_"`.
## Note that only values of `col1` for rows matching at least one query
## row are returned and an NA is reported for query rows without matching
## target rows.
mo$target_col1
#> [1] 12 12 13 14 NA NA 15

## The 3rd and 4th query rows do not match any target row, thus `NA` is
## returned.

## `matchedData` can be used to extract all (or selected) columns
## from the object. Same as with `$`, a left join between the columns
## from the query and the target is performed. Below we extract selected
## columns from the object as a DataFrame.
res <- matchedData(mo, columns = c("col1", "col2", "target_col1", "target_col2"))
res
#> DataFrame with 7 rows and 4 columns
#>          col1      col2 target_col1 target_col2
#>     <integer> <integer>   <integer>   <integer>
#> 1           1         6          12          18
#> 2           2         7          12          18
#> 2.1         2         7          13          19
#> 2.2         2         7          14          20
#> 3           3         8          NA          NA
#> 4           4         9          NA          NA
#> 5           5        10          15          21
res$col1
#> [1] 1 2 2 2 3 4 5
res$target_col1
#> [1] 12 12 13 14 NA NA 15

## The example matched object contains all query and all target
## elements (rows). Below we subset the object keeping only query rows that
## are matched to at least one target row.
mo_sub <- mo[whichQuery(mo)]

## ms_sub contains now only 3 query rows:
nrow(query(mo_sub))
#> [1] 3

## while the original object contains all 5 query rows:
nrow(query(mo))
#> [1] 5

## Both object contain however still the full target object:
nrow(target(mo))
#> [1] 6
nrow(target(mo_sub))
#> [1] 6

## With the `pruneTarget` we can however reduce also the target rows to
## only those that match at least one query row
mo_sub <- pruneTarget(mo_sub)
nrow(target(mo_sub))
#> [1] 4

## Creating a Matched object with a `data.frame` for `query` and a `vector`
## for `target`. The matches are specified in the same way as the example
## before.

q1 <- data.frame(col1 = 1:5, col2 = 6:10)
t2 <- 11:16
mo <- Matched(q1, t2, matches = data.frame(query_idx = c(1L, 2L, 2L, 2L, 5L),
    target_idx = c(2L, 2L, 3L, 4L, 5L), score = seq(0.5, 0.9, by = 0.1)))

## *target* is a simple vector and has thus no columns. The matched values
## from target, if it does not have dimensions and hence column names, can
## be retrieved with `$target`
mo$target
#> [1] 12 12 13 14 NA NA 15

## Note that in this case "target" is returned by the function `colnames`
colnames(mo)
#> [1] "col1"   "col2"   "target" "score" 

## As before, we can extract all data as a `DataFrame`
res <- matchedData(mo)
res
#> DataFrame with 7 rows and 4 columns
#>          col1      col2    target     score
#>     <integer> <integer> <integer> <numeric>
#> 1           1         6        12       0.5
#> 2           2         7        12       0.6
#> 2.1         2         7        13       0.7
#> 2.2         2         7        14       0.8
#> 3           3         8        NA        NA
#> 4           4         9        NA        NA
#> 5           5        10        15       0.9

## Note that the columns of the obtained `DataFrame` are the same as the
## corresponding vectors obtained with `$`

res$col1
#> [1] 1 2 2 2 3 4 5
res$target
#> [1] 12 12 13 14 NA NA 15

## Also subsetting and pruning works in the same way as the example above.

mo_sub <- mo[whichQuery(mo)]

## mo_sub contains now only 3 query rows:
nrow(query(mo_sub))
#> [1] 3

## while the original object contains all 5 query rows:
nrow(query(mo))
#> [1] 5

## Both object contain however still the full target object:
length(target(mo))
#> [1] 6
length(target(mo_sub))
#> [1] 6

## Reducing the target elements to only those that match at least one query row
mo_sub <- pruneTarget(mo_sub)
nrow(target(mo_sub))
#> NULL