m/z matching

The matchMz method matches elements from query with those in target using different matching approaches depending on parameter param. Generally, query is expected to contain MS experimental values (m/z and possibly retention time) while target theoretical values from reference compounds. target can be numeric or data.frameOrSimilar (i.e. data.frame, matrix or DataFrame). The same is true for query which in addition can also be SummarizedExperiment.

Available param objects and corresponding matching approaches are:

Mass2MzParam: match m/z values against reference compounds for which the (exact) mass is known. Before matching, m/z values are calculated from the compounds masses in the target table using the adducts specified via Mass2MzParam adducts parameter (defaults to adducts = "[M+H]+"). query must be a data.frameOrSimilar with a column containing m/z values (column name can be specified with parameter mzColname which defaults to "mz") or a numeric with the m/z values of the features. target must be a data.frameOrSimilar with a column containing (monoisotopic) masses of the reference compounds (the column name can be specified with parameter massColname which defaults to "exactmass") or a numeric with the mass values. Parameters tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between m/zs in query and those obtained from masses in target.

Mass2MzRtParam: match m/z and retention time values against reference compounds for which the (exact) mass and retention time are known. Before matching, m/z values are calculated from the compounds masses in the target table using the adducts specified via Mass2MzParam adducts parameter (defaults to adducts = "[M+H]+"). The retention time is considered equal for different adducts of the same compound. query must be a data.frameOrSimilar with m/z and retention time values of the features. The names of the columns containing these information can be specified with parameters mzColname and rtColname which default to "mz" and "rt", respectively. target must be a data.frameOrSimilar with the (monoisotopic) masses and retention times for each reference compound. The names of the columns containing this information can be specified with parameters massColname and rtColname which default to "exactmass" and "rt", respectively. Mass2MzRtParam parameters tolerance and ppm have the same meaning as in Mass2MzParam; parameter toleranceRt allows to specify the maximal acceptable difference between query and target retention time values.

MzParam: match m/z values against reference compounds for which the m/z is known. query must be either a data.frameOrSimilar with a column containing m/z values (whose name can be specified with parameter mzColname, default being mzColname = "mz") or a numeric with the m/z values of the features. The same holds for target. MzParam parameters tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values.

MzRtParam: match m/z and retention time values against reference compounds for which m/z and retention time are known. query must be a data.frameOrSimilar or a SummarizedExperiment query in the first case or its rowData in the second must have columns containing the m/zs and retention times of the features. target must be a data.frameOrSimilar again with columns containing m/zs and retention times. The names of such columns can be specified with parameters mzColname and rtColname (defaulting to c("mz", "mz") and c("rt", "rt"), respectively). MzRtParam parameters tolerance and ppm have the same meaning as in MzParam; MzRtParam parameter toleranceRt allows to specify the maximal acceptable difference between query and target retention time values.

Mz2MassParam: first convert to masses the m/z values of query and target based respectively on Mz2MassParam parameters queryAdducts (defaults to queryAdducts = "[M+H]+") and targetAdducts (defaults to targetAdducts = "[M-H]-") and then match the obtained masses. query must be either a data.frameOrSimilar with a column containing m/z values (whose name can be specified with parameter mzColname, default being mzColname = "mz") or a numeric with the m/z values. The same holds for target. Parameters tolerance and ppm allow to define the maximal acceptable (constant or mass relative) difference between masses derived from query and target.

Mz2MassRtParam: first convert to masses the m/z values of query and target based respectively on Mz2MassRtParam parameters queryAdducts (defaults to queryAdducts = "[M+H]+") and targetAdducts (defaults to targetAdducts = "[M-H]-") and then match the obtained masses and corresponding retention times (to each mass value corresponds the same retention time associated to the m/z from which the mass value was obtained). query must be a data.frameOrSimilar with a column containing m/z values (whose name can be specified with parameter mzColname, default being mzColname = "mz"). The same holds for target. Parameters tolerance and ppm have the same meaning as in Mz2MassParam. Parameter toleranceRt allows to specify the maximal acceptable difference between retention time values.

ValueParam: matches elements from from query (if numeric) or else from one of its columns with those from target (if numeric) or else from one of its columns. The name(s) of the colum(s) used for the matching (if query and target are not bot numeric has/have to be specified via parameter valueColname.

Usage

ValueParam(tolerance = 0, ppm = 5)

MzParam(tolerance = 0, ppm = 5)

Mass2MzParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5)

Mass2MzRtParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5, toleranceRt = 0)

MzRtParam(tolerance = 0, ppm = 0, toleranceRt = 0)

Mz2MassParam(
  queryAdducts = c("[M+H]+"),
  targetAdducts = c("[M-H]-"),
  tolerance = 0,
  ppm = 5
)

Mz2MassRtParam(
  queryAdducts = c("[M+H]+"),
  targetAdducts = c("[M+H]+"),
  tolerance = 0,
  ppm = 5,
  toleranceRt = 0
)

matchMz(query, target, param, ...)

# S4 method for numeric,numeric,ValueParam
matchMz(query, target, param)

# S4 method for numeric,data.frameOrSimilar,ValueParam
matchMz(query, target, param, valueColname = character())

# S4 method for data.frameOrSimilar,numeric,ValueParam
matchMz(query, target, param, valueColname = character())

# S4 method for data.frameOrSimilar,data.frameOrSimilar,ValueParam
matchMz(query, target, param, valueColname = character())

# S4 method for numeric,numeric,Mass2MzParam
matchMz(query, target, param)

# S4 method for numeric,data.frameOrSimilar,Mass2MzParam
matchMz(query, target, param, massColname = "exactmass")

# S4 method for data.frameOrSimilar,numeric,Mass2MzParam
matchMz(query, target, param, mzColname = "mz")

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mass2MzParam
matchMz(query, target, param, mzColname = "mz", massColname = "exactmass")

# S4 method for numeric,data.frameOrSimilar,MzParam
matchMz(query, target, param, mzColname = "mz")

# S4 method for data.frameOrSimilar,numeric,MzParam
matchMz(query, target, param, mzColname = "mz")

# S4 method for data.frameOrSimilar,data.frameOrSimilar,MzParam
matchMz(query, target, param, mzColname = c("mz", "mz"))

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mass2MzRtParam
matchMz(
  query,
  target,
  param,
  massColname = "exactmass",
  mzColname = "mz",
  rtColname = c("rt", "rt")
)

# S4 method for data.frameOrSimilar,data.frameOrSimilar,MzRtParam
matchMz(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  rtColname = c("rt", "rt")
)

# S4 method for numeric,numeric,Mz2MassParam
matchMz(query, target, param)

# S4 method for numeric,data.frameOrSimilar,Mz2MassParam
matchMz(query, target, param, mzColname = "mz")

# S4 method for data.frameOrSimilar,numeric,Mz2MassParam
matchMz(query, target, param, mzColname = "mz")

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mz2MassParam
matchMz(query, target, param, mzColname = c("mz", "mz"))

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mz2MassRtParam
matchMz(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  rtColname = c("rt", "rt")
)

# S4 method for SummarizedExperiment,ANY,Param
matchMz(query, target, param, mzColname = "mz", rtColname = c("rt", "rt"))

Arguments

tolerance

for any param object: numeric(1) defining the maximal acceptable absolute difference in m/z (or in mass for Mz2MassParam) to consider them matching.

ppm

for any param object: numeric(1) defining the maximal acceptable m/z-dependent (or mass-dependent for Mz2MassParam) difference (in parts-per-million) in m/z values to consider them to be matching.

adducts

for Mass2MzParam or Mass2MzRtParam: either character with adduct names from MetaboCoreUtils::adducts() or data.frame with a custom adduct definition. This parameter is used to calculate m/z from target compounds' masses. Custom adduct definitions can be passed to the adduct parameter in form of a data.frame. This data.frame is expected to have columns "mass_add" and "mass_multi" defining the additive and multiplicative part of the calculation. See MetaboCoreUtils::adducts() for the expected format or use MetaboCoreUtils::adductNames("positive") and MetaboCoreUtils::adductNames("negative") for valid adduct names.

toleranceRt

for Mass2MzRtParam or MzRtParam: numeric(1) defining the maximal acceptable absolute difference in retention time values to consider them them matching.

queryAdducts

for Mz2MassParam. Adducts used to derive mass values from query m/z values. The expected format is the same as that for parameter adducts.

targetAdducts

for Mz2MassParam. Adducts used to derive mass values from target m/z values. The expected format is the same as that for parameter adducts.

query

feature table containing information on MS1 features. Can be a data.frame, DataFrame, matrix, SummarizedExperiment or numeric. It is expected to contain m/z values and can contain also other variables. Matchings based on both m/z and retention time can be performed when a column with retention times is present in both query and target.

target

compound table with metabolites to compare against.

param

parameter object defining the matching approach and containing the settings for that approach. See description above for details.

...

currently ignored.

valueColname

character specifying the name of the column in query or/and the one in targetwith the desired values for the matching. This parameter should only be used when param is valueParam and in this case it must be provided (unless both query and target are numeric). It can be character(1) or character(2) in a similar way to mzColname.

massColname

character(1) with the name of the column in target containing the mass of compounds. To be used when param is Mass2MzParam or Mass2MzRtParam (and target is not already numeric with the masses). Defaults to massColname = "exactmass".

mzColname

character specifying the name(s) of the column(s) in query or/and targetwith the m/z values. If one among query and target is numeric (and therefore there is no need to specify the column name) or query is not numeric and param is Mass2MzParam or Mass2MzRtParam (and therefore the name of the column with m/z needs only to be specified for query) then mzColname is expected to be character(1). If both query and target are not numeric mzColname is expected to be character(2) (or character(1) and in this last case the two column names are assumed to be the same). If not specified the assumed default name for coulmns with m/z values is "mz".

rtColname

character(2) with the name of the column containing the compounds retention times in query and the name for the one in target. It can also be character(1) if the two names are the same. To be used when param is MzRtParam or Mass2MzRtParam. Defaults to rtColname = c("rt", "rt").

Value

Matched object representing the result (or more precisely MatchedSummarizedExperiment if query is SummarizedExperiment). To evaluate each match the object contains, depending on which param is used, the m/z or mass error in ppm (variable "ppm_error") as well as the m/z or mass difference (variable "score"). When param is Mz2MassParam such difference and ppm error are obtained using mass values internally computed from m/z values in query and target (see above for more detail on Mz2MassParam). When parm is either MzParam, MzRtParam, Mass2MzParam or Mass2MzRtParam"ppm_error" and "score"are computed from m/z values. In the first two cases both the m/z values are contained in query and target parameters. In the last two m/z values are internally computed from reference mass values in target. See more details for each individual parameter above. Additionally, if param is either MzRtParam or Mass2MzRtParam also retention time is used for the matching and the difference in retention time for each of the matched elements is also added to the returned Matched object. Note that, for a match, a negative value of "score" (or "score_rt") indicates that the m/z or mass (or retention time) associated to the query element is smaller than that associated to the target element.

See also

Author

Andrea Vicini, Michael Witting

Examples

library(MetaboCoreUtils)
## Create a simple "target/reference" compound table
target_df <- data.frame(
   name = c("Tryptophan", "Leucine", "Isoleucine"),
   formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
   exactmass = c(204.089878, 131.094629, 131.094629)
)

## Create a "feature" table with m/z of features. We calculate m/z for
## certain adducts of some of the compounds in the reference table.
fts <- data.frame(
    feature_id = c("FT001", "FT002", "FT003"),
    mz = c(mass2mz(204.089878, "[M+H]+"),
           mass2mz(131.094629, "[M+H]+"),
           mass2mz(204.089878, "[M+Na]+") + 1e-6))

## Define the parameters for the matching
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 4 
#> Number of query objects: 3 (3 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 8 columns
#>      feature_id        mz target_name target_formula target_exactmass
#>     <character> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079  Tryptophan     C11H12N2O2          204.090
#>          adduct     score  ppm_error
#>     <character> <numeric>  <numeric>
#> 1        [M+H]+     0e+00 0.00000000
#> 2        [M+H]+     0e+00 0.00000000
#> 2.1      [M+H]+     0e+00 0.00000000
#> 3       [M+Na]+     1e-06 0.00440375

## We have thus matches of FT002 to two different compounds (but with the
## same mass).

## We repeat the matching requiring an exact match
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 0)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 3 
#> Number of query objects: 3 (2 matched)
#> Number of target objects: 3 (3 matched)

matchedData(res)
#> DataFrame with 4 rows and 8 columns
#>      feature_id        mz target_name target_formula target_exactmass
#>     <character> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079          NA             NA               NA
#>          adduct     score ppm_error
#>     <character> <numeric> <numeric>
#> 1        [M+H]+         0         0
#> 2        [M+H]+         0         0
#> 2.1      [M+H]+         0         0
#> 3            NA        NA        NA

## The last feature could thus not be matched to any compound.

## At last we use also different adduct definitions.
parm <- Mass2MzParam(
    adducts = c("[M+K]+", "[M+Li]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 0 
#> Number of query objects: 3 (0 matched)
#> Number of target objects: 3 (0 matched)

matchedData(res)
#> DataFrame with 3 rows and 8 columns
#>    feature_id        mz target_name target_formula target_exactmass      adduct
#>   <character> <numeric> <character>    <character>        <numeric> <character>
#> 1       FT001   205.097          NA             NA               NA          NA
#> 2       FT002   132.102          NA             NA               NA          NA
#> 3       FT003   227.079          NA             NA               NA          NA
#>       score ppm_error
#>   <numeric> <numeric>
#> 1        NA        NA
#> 2        NA        NA
#> 3        NA        NA

## No matches were found.

## We can also match a "feature" table with a target data.frame taking into
## account both m/z and retention time values.
target_df <- data.frame(
  name = c("Tryptophan", "Leucine", "Isoleucine"),
  formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
  exactmass = c(204.089878, 131.094629, 131.094629),
  rt = c(150, 140, 140)
)

fts <- data.frame(
  feature_id = c("FT001", "FT002", "FT003"),
  mz = c(mass2mz(204.089878, "[M+H]+"),
         mass2mz(131.094629, "[M+H]+"),
         mass2mz(204.089878, "[M+Na]+") + 1e-6),
  rt = c(150, 140, 150.1)
)

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0)

res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 3 
#> Number of query objects: 3 (2 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 11 columns
#>      feature_id        mz        rt target_name target_formula target_exactmass
#>     <character> <numeric> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097     150.0  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     140.0     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102     140.0  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079     150.1          NA             NA               NA
#>     target_rt      adduct     score ppm_error  score_rt
#>     <numeric> <character> <numeric> <numeric> <numeric>
#> 1         150      [M+H]+         0         0         0
#> 2         140      [M+H]+         0         0         0
#> 2.1       140      [M+H]+         0         0         0
#> 3          NA          NA        NA        NA        NA

## FT003 could not be matched to any compound, FT002 was matched to two
## different compounds (but with the same mass).

## We repeat the matching allowing a positive tolerance for the matches
## between rt values

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0.1)

res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 4 
#> Number of query objects: 3 (3 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 11 columns
#>      feature_id        mz        rt target_name target_formula target_exactmass
#>     <character> <numeric> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097     150.0  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     140.0     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102     140.0  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079     150.1  Tryptophan     C11H12N2O2          204.090
#>     target_rt      adduct     score  ppm_error  score_rt
#>     <numeric> <character> <numeric>  <numeric> <numeric>
#> 1         150      [M+H]+     0e+00 0.00000000       0.0
#> 2         140      [M+H]+     0e+00 0.00000000       0.0
#> 2.1       140      [M+H]+     0e+00 0.00000000       0.0
#> 3         150     [M+Na]+     1e-06 0.00440375       0.1

## Also FT003 was matched in this case

## It is also possible to match directly m/z values
mz1 <- c(12, 343, 23, 231)
mz2 <- mz1 + rnorm(4, sd = 0.001)

res <- matchMz(mz1, mz2, MzParam(tolerance = 0.001))

matchedData(res)
#> DataFrame with 4 rows and 4 columns
#>     query  target        score ppm_error
#>    <AsIs>  <AsIs>    <numeric> <numeric>
#> NA     12      NA           NA        NA
#> 1     343 343.000 -1.14620e-04   0.33417
#> 2      23  23.000  2.59175e-05   1.12685
#> 3     231 231.001 -1.24756e-03   5.40068