Spectral matching

The matchSpectra method matches (compares) spectra from query with those from target based on settings specified with param and returns the result from this as a MatchedSpectra object.

Usage

matchSpectra(query, target, param, ...)

Arguments

query

Spectra object with the (experimental) spectra.

target

spectral data to compare against. Can be another Spectra.

param

parameter object containing the settings for the matching (e.g. eventual prefiltering settings, cut-off value for similarity above which spectra are considered matching etc).

...

optional parameters.

Value

a MatchedSpectra object with the spectra matching results.

See also

Author

Johannes Rainer