m/z matching

The matchMz method matches (compares) m/z values from a MS1 data table (parameter query) with theoretical m/z values for (reference) compounds (parameter target). The approach which is used to perform the comparison and additional settings for this matching can be defined with param.

Available matching approaches and respective param objects are:

TargetMass2MzParam: match m/z values against reference compounds for which the (exact) mass is known. Before matching, m/z values are calculated from the mass of the compounds in the target table for the specified adducts (parameter adduct in the parameter object). query must be a data.frame with a column named "mz" or a numeric with the m/z values of the features. If target is a data.frame it must contain a column "exactmass" with the exact monoisotopic mass for each compound. TargetMass2MzParam's parameter adducts allows to define the expected adducts (defaults to adducts = "[M+H]+" but any adducts available in [MetaboCoreUtils::adducts()] are supported). Parameter toleranceandppm` allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values.

TargetMass2MzParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5)

matchMz(query, target, param, ...)

# S4 method for numeric,data.frame,TargetMass2MzParam
matchMz(query, target, param, BPPARAM = SerialParam())

# S4 method for data.frame,data.frame,TargetMass2MzParam
matchMz(query, target, param, BPPARAM = SerialParam())

Arguments

adducts
for TargetMass2MzParam: character with the names of adducts to calculate m/z from target compounds' masses. Use MetaboCoreUtils::adductNames("positive") and MetaboCoreUtils::adductNames("negative") for valid names.

tolerance
for TargetMass2MzParam: numeric(1) defining the maximal acceptable absolute difference in m/z values to consider them matching. See MsCoreUtils::closest() for more information.

ppm
for TargetMass2MzParam: numeric(1) defining the maximal acceptable m/z-dependent difference (in parts-per-million) in m/z values to consider them to be matching. See MsCoreUtils::closest() for more information.

query
feature table containing information on MS1 features. Can be a data.frame (with mandatory column names "mz") or a numeric with the m/z values.

target
compound table with metabolites to compare against.

param
parameter object defining the matching approach and containing the settings for that approach. See description above for details.

...
currently ignored.

BPPARAM
parallel processing setup. See BiocParallel::bpparam() for details.

Value

Matched object representing the result

See also

Author

Andrea Vicini, Michael Witting

Examples

library(MetaboCoreUtils)
## Create a simple "target/reference" compound table
target_df <- data.frame(
   name = c("Tryptophan", "Leucine", "Isoleucine"),
   formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
   exactmass = c(204.089878, 131.094629, 131.094629)
)

## Create a "feature" table with m/z of features. We calculate m/z for
## certain adducts of some of the compounds in the reference table.
fts <- data.frame(
    feature_id = c("FT001", "FT002", "FT003"),
    mz = c(mass2mz(204.089878, "[M+H]+"),
           mass2mz(131.094629, "[M+H]+"),
           mass2mz(204.089878, "[M+Na]+") + 1e-6))

## Define the parameters for the matching
parm <- TargetMass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 4 
#> Number of query objects: 3 (3 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
data(res)
#> Warning: data set ‘res’ not found

## We have thus matches of FT002 to two different compounds (but with the
## same mass.

## We repeat the matching requiring an exact match
parm <- TargetMass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 0)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 3 
#> Number of query objects: 3 (2 matched)
#> Number of target objects: 3 (3 matched)

data(res)
#> Warning: data set ‘res’ not found

## The last feature could thus not be matched to any compound.

## At last we use also different adduct definitions.
parm <- TargetMass2MzParam(
    adducts = c("[M+K]+", "[M+Li]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 0 
#> Number of query objects: 3 (0 matched)
#> Number of target objects: 3 (0 matched)

data(res)
#> Warning: data set ‘res’ not found

## No matches were found.