m/z matching

The matchMz method matches (compares) m/z values from a MS1 data table (parameter query) with theoretical m/z values for (reference) compounds (parameter target) considering or not also retention times. The approach which is used to perform the comparison and additional settings for this matching can be defined with param.

Available matching approaches and respective param objects are:

Mass2MzParam: match m/z values against reference compounds for which the (exact) mass is known. Before matching, m/z values are calculated from the mass of the compounds in the target table for the specified adducts (parameter adduct in the parameter object). query must be a data.frame with a column containing m/z values (column name can be specified with parameter mzColname which defaults to "mz") or a numeric with the m/z values of the features. If target is a data.frame it must contain a column with the (monoisotopic) mass of the compounds (the column name can be specified with parameter massColname which defaults to "exactmass") Mass2MzParam's parameter adducts allows to define the expected adducts (defaults to adducts = "[M+H]+" but any adducts available in MetaboCoreUtils::adducts() are supported). Parameter tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values.

Mass2MzRtParam: match m/z and retention time values against reference compounds for which the (exact) mass and retention time are known. Before matching, m/z values are calculated from the mass of the compounds in the target table for the specified adducts (parameter adduct in the parameter object). The retention time is considered equal for different adducts of the same compound. query must be a data.frame with m/z and retention time values of the features. The names of the columns containing these information can be specified with parameters mzColname and rtColname which default to "mz" and "rt", respectively). target must be a data.frame with the (monoisotopic) mass and retention time for each compound. The names of the columns containing these information can be specified with parameters massColname and rtColname which default to "exactmass" and "rt", respectively. Mass2MzRtParam's parameter adducts allows to define the expected adducts (defaults to adducts = "[M+H]+" but any adducts available in MetaboCoreUtils::adducts() are supported). Parameter tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values; parameter toleranceRt allows to specify the maximal acceptable difference between query and target retention time values.

MzParam: match m/z values against reference compounds for which the m/z is known. query must be either a data.frame with a column containing m/z values (which name can be specified with parameter mzColname, default being mzColname = "mz") or a numeric with the m/z values of the features. The same holds for target. MzParam parameters tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values.

MzRtParam: match m/z and retention time values against reference compounds for which m/z and retention time are known. query must be a data.frame or a SummarizedExperiment. The data.frame in one case or the SummarizedExperiment rowData in the other must have columns containing the m/z and retention times of the features. The names of the respective columns can be specified with parameters mzColname and rtColname which default to "mz" and "rt", respectively.target must be a data.frame again with the information on m/z and retention time.MzRtParam parameters tolerance and ppm allow to define the maximal acceptable (constant or m/z relative) difference between query and target m/z values; MzRtParam parameter toleranceRt allows to specify the maximal acceptable difference between query and target retention time values.

Usage

MzParam(tolerance = 0, ppm = 5)

Mass2MzParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5)

Mass2MzRtParam(adducts = c("[M+H]+"), tolerance = 0, ppm = 5, toleranceRt = 0)

MzRtParam(tolerance = 0, ppm = 0, toleranceRt = 0)

matchMz(query, target, param, ...)

# S4 method for numeric,numeric,Mass2MzParam
matchMz(query, target, param, BPPARAM = SerialParam())

# S4 method for numeric,data.frameOrSimilar,Mass2MzParam
matchMz(
  query,
  target,
  param,
  massColname = "exactmass",
  BPPARAM = SerialParam()
)

# S4 method for data.frameOrSimilar,numeric,Mass2MzParam
matchMz(query, target, param, BPPARAM = SerialParam(), mzColname = "mz")

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mass2MzParam
matchMz(
  query,
  target,
  param,
  mzColname = "mz",
  massColname = "exactmass",
  BPPARAM = SerialParam()
)

# S4 method for numeric,numeric,MzParam
matchMz(query, target, param, BPPARAM = SerialParam())

# S4 method for numeric,data.frameOrSimilar,MzParam
matchMz(query, target, param, mzColname = "mz", BPPARAM = SerialParam())

# S4 method for data.frameOrSimilar,numeric,MzParam
matchMz(query, target, param, mzColname = "mz", BPPARAM = SerialParam())

# S4 method for data.frameOrSimilar,data.frameOrSimilar,MzParam
matchMz(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  BPPARAM = SerialParam()
)

# S4 method for data.frameOrSimilar,data.frameOrSimilar,Mass2MzRtParam
matchMz(
  query,
  target,
  param,
  massColname = "exactmass",
  mzColname = "mz",
  rtColname = c("rt", "rt"),
  BPPARAM = SerialParam()
)

# S4 method for data.frameOrSimilar,data.frameOrSimilar,MzRtParam
matchMz(
  query,
  target,
  param,
  mzColname = c("mz", "mz"),
  rtColname = c("rt", "rt"),
  BPPARAM = SerialParam()
)

# S4 method for SummarizedExperiment,ANY,Param
matchMz(
  query,
  target,
  param,
  mzColname = "mz",
  rtColname = "rt",
  BPPARAM = SerialParam()
)

Arguments

tolerance

for any param object: numeric(1) defining the maximal acceptable absolute difference in m/z values to consider them matching.

ppm

for any param object: numeric(1) defining the maximal acceptable m/z-dependent difference (in parts-per-million) in m/z values to consider them to be matching.

adducts

for Mass2MzParam or Mass2MzRtParam: either character with the names of adducts or data.frame with the adduct definition. This parameter is used to calculate m/z from target compounds' masses. Custom adduct definitions can be passed to the adduct parameter in form of a data.frame. This data.frame is expected to have columns "mass_add" and "mass_multi" defining the additive and multiplicative part of the calculation. See MetaboCoreUtils::adducts() for the expected format or use MetaboCoreUtils::adductNames("positive") and MetaboCoreUtils::adductNames("negative") for valid adduct names.

toleranceRt

for Mass2MzRtParam or MzRtParam: numeric(1) defining the maximal acceptable absolute difference in retention time values to consider them them matching.

query

feature table containing information on MS1 features. Can be a data.frame (with mandatory column names "mz") or a numeric with the m/z values. A matching based on both m/z and retention time can be performed when a column "rt" is present in both query and target.

target

compound table with metabolites to compare against.

param

parameter object defining the matching approach and containing the settings for that approach. See description above for details.

...

currently ignored.

BPPARAM

parallel processing setup. See BiocParallel::bpparam() for details.

massColname

character(1) with the name of the column containing the mass of compounds. Defaults to massColname = "exactmass".

mzColname

character(1) with the name of the column containing the m/z values. Defaults to mzColname = "mz".

rtColname

character(1) with the name of the column containing the retention times of the compounds. Defaults to rtColname = "rt".

Value

Matched object representing the result. To evaluate each match the object contains the m/z error in ppm (variable "ppm_error") as well as the difference between the query and target m/z (variable "score"). The difference between the query and target retention time (variable "score_rt"is also present if retention time is considered for the match. Thus, for a match, a negative value of "score" (or "score_rt") indicates that the m/z (or retention time) of the query element is smaller than that in the matched target element.

See also

Author

Andrea Vicini, Michael Witting

Examples

library(MetaboCoreUtils)
## Create a simple "target/reference" compound table
target_df <- data.frame(
   name = c("Tryptophan", "Leucine", "Isoleucine"),
   formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
   exactmass = c(204.089878, 131.094629, 131.094629)
)

## Create a "feature" table with m/z of features. We calculate m/z for
## certain adducts of some of the compounds in the reference table.
fts <- data.frame(
    feature_id = c("FT001", "FT002", "FT003"),
    mz = c(mass2mz(204.089878, "[M+H]+"),
           mass2mz(131.094629, "[M+H]+"),
           mass2mz(204.089878, "[M+Na]+") + 1e-6))

## Define the parameters for the matching
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 4 
#> Number of query objects: 3 (3 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 8 columns
#>      feature_id        mz target_name target_formula target_exactmass
#>     <character> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079  Tryptophan     C11H12N2O2          204.090
#>          adduct     score  ppm_error
#>     <character> <numeric>  <numeric>
#> 1        [M+H]+     0e+00 0.00000000
#> 2        [M+H]+     0e+00 0.00000000
#> 2.1      [M+H]+     0e+00 0.00000000
#> 3       [M+Na]+     1e-06 0.00440375

## We have thus matches of FT002 to two different compounds (but with the
## same mass).

## We repeat the matching requiring an exact match
parm <- Mass2MzParam(
    adducts = c("[M+H]+", "[M+Na]+"),
    tolerance = 0,
    ppm = 0)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 3 
#> Number of query objects: 3 (2 matched)
#> Number of target objects: 3 (3 matched)

matchedData(res)
#> DataFrame with 4 rows and 8 columns
#>      feature_id        mz target_name target_formula target_exactmass
#>     <character> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079          NA             NA               NA
#>          adduct     score ppm_error
#>     <character> <numeric> <numeric>
#> 1        [M+H]+         0         0
#> 2        [M+H]+         0         0
#> 2.1      [M+H]+         0         0
#> 3            NA        NA        NA

## The last feature could thus not be matched to any compound.

## At last we use also different adduct definitions.
parm <- Mass2MzParam(
    adducts = c("[M+K]+", "[M+Li]+"),
    tolerance = 0,
    ppm = 20)
res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 0 
#> Number of query objects: 3 (0 matched)
#> Number of target objects: 3 (0 matched)

matchedData(res)
#> DataFrame with 3 rows and 8 columns
#>    feature_id        mz target_name target_formula target_exactmass      adduct
#>   <character> <numeric> <character>    <character>        <numeric> <character>
#> 1       FT001   205.097          NA             NA               NA          NA
#> 2       FT002   132.102          NA             NA               NA          NA
#> 3       FT003   227.079          NA             NA               NA          NA
#>       score       ppm
#>   <numeric> <numeric>
#> 1        NA        NA
#> 2        NA        NA
#> 3        NA        NA

## No matches were found.

## We can also match a "feature" table with a target data.frame taking into
## account both m/z and retention time values.
target_df <- data.frame(
  name = c("Tryptophan", "Leucine", "Isoleucine"),
  formula = c("C11H12N2O2", "C6H13NO2", "C6H13NO2"),
  exactmass = c(204.089878, 131.094629, 131.094629),
  rt = c(150, 140, 140)
)

fts <- data.frame(
  feature_id = c("FT001", "FT002", "FT003"),
  mz = c(mass2mz(204.089878, "[M+H]+"),
         mass2mz(131.094629, "[M+H]+"),
         mass2mz(204.089878, "[M+Na]+") + 1e-6),
  rt = c(150, 140, 150.1)
)

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0)

res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 3 
#> Number of query objects: 3 (2 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 11 columns
#>      feature_id        mz        rt target_name target_formula target_exactmass
#>     <character> <numeric> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097     150.0  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     140.0     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102     140.0  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079     150.1          NA             NA               NA
#>     target_rt      adduct     score ppm_error  score_rt
#>     <numeric> <character> <numeric> <numeric> <numeric>
#> 1         150      [M+H]+         0         0         0
#> 2         140      [M+H]+         0         0         0
#> 2.1       140      [M+H]+         0         0         0
#> 3          NA          NA        NA        NA        NA

## FT003 could not be matched to any compound, FT002 was matched to two
## different compounds (but with the same mass).

## We repeat the matching allowing a positive tolerance for the matches
## between rt values

## Define the parameters for the matching
parm <- Mass2MzRtParam(
  adducts = c("[M+H]+", "[M+Na]+"),
  tolerance = 0,
  ppm = 20,
  toleranceRt = 0.1)

res <- matchMz(fts, target_df, parm)
res
#> Object of class Matched 
#> Total number of matches: 4 
#> Number of query objects: 3 (3 matched)
#> Number of target objects: 3 (3 matched)

## Get the full matching result:
matchedData(res)
#> DataFrame with 4 rows and 11 columns
#>      feature_id        mz        rt target_name target_formula target_exactmass
#>     <character> <numeric> <numeric> <character>    <character>        <numeric>
#> 1         FT001   205.097     150.0  Tryptophan     C11H12N2O2          204.090
#> 2         FT002   132.102     140.0     Leucine       C6H13NO2          131.095
#> 2.1       FT002   132.102     140.0  Isoleucine       C6H13NO2          131.095
#> 3         FT003   227.079     150.1  Tryptophan     C11H12N2O2          204.090
#>     target_rt      adduct     score  ppm_error  score_rt
#>     <numeric> <character> <numeric>  <numeric> <numeric>
#> 1         150      [M+H]+     0e+00 0.00000000       0.0
#> 2         140      [M+H]+     0e+00 0.00000000       0.0
#> 2.1       140      [M+H]+     0e+00 0.00000000       0.0
#> 3         150     [M+Na]+     1e-06 0.00440375       0.1

## Also FT003 was matched in this case

## It is also possible to match directly m/z values
mz1 <- c(12, 343, 23, 231)
mz2 <- mz1 + rnorm(4, sd = 0.001)

res <- matchMz(mz1, mz2, MzParam(tolerance = 0.001))

matchedData(res)
#> DataFrame with 4 rows and 4 columns
#>        query    target        score  ppm_error
#>    <numeric> <numeric>    <numeric>  <numeric>
#> NA        12        NA           NA         NA
#> 1        343  343.0002 -1.94151e-04  0.5660371
#> 2         23   22.9994  5.76006e-04 25.0443754
#> 3        231  231.0000  2.36293e-06  0.0102291